feat(impl):[TRI-1610] dont validate level zero

+++ irs-models/src/main/java/org/eclipse/tractusx/irs/component/enums/IntegrityState.java
@@ -28,7 +28,7 @@ package org.eclipse.tractusx.irs.component.enums;
 public enum IntegrityState {
     VALID, INVALID, INACTIVE;
 
-    public static IntegrityState from(final long numberOfValidSubmodels, final int totalNumberOfSubmodels) {
+    public static IntegrityState from(final long numberOfValidSubmodels, final long totalNumberOfSubmodels) {
         if (numberOfValidSubmodels == totalNumberOfSubmodels) {
             return VALID;
         } else {
